// src/lib/d4h-api/whoami-response.d.ts
export interface WhoamiResponse {

    account: {
        id: number,
        resourceType: string
    },
    members: {
        hasAccess: boolean,
        id: number,
        name: string,
        owner: {
            id: number
            resourceType: string,
            title: string
        },
        permissions?: Record<
            'Animal' | 'AnimalGroup' | 'AnimalQualification' | 'AnimalQualificationAward' | 'Audit' | 'BillingUnit' | 'CustomField' | 'CustomFieldOption' | 'CustomIdentifier' |
            'Division' | 'Document' | 'Duty' | 'Equipment' | 'EquipmentInspection' | 'EquipmentInspectionResult' | 'EquipmentLocation' | 'EquipmentUsage' | 'Event' | 'Exercise' |
            'HandlerGroup' | 'HandlerQualification' | 'HandlerQualificationAward' | 'HealthAndSafetyCategory' | 'HealthAndSafetyReport' | 'HealthAndSafetySeverity' | 'Incident' |
            'LocationBookmark' | 'Member' | 'MemberGroup' | 'MemberQualification' | 'MemberQualificationAward' | 'PersonInvolved' | 'Repair' | 'Resource' | 'ResourceBundle' |
            'Role' | 'Setting' | 'Tag' | 'Task' | 'Team' | 'Whiteboard'
            , 
            Record<PermissionType, boolean | undefined>>
    }[]
};

type PermissionType = 'APPROVE' | 'ARCHIVED' | 'ASSIGN_UNASSIGED' | 'CREATE' | 'CREATE_CUSTOM_FIELDS' | 'CREATE_SECURE' | 'DEFAULT_ACCESS' | 'DELETE' | 'DELETE_CUSTOM_FIELD' | 'DELETE_SECURE' | 'DOWNLOAD_SECURE' | 'DRAFT' | 'EXPORT' | 'LIST' | 'READ' | 'READ_SECURE' | 'UPDATE' | 'UPDATE_CUSTOM_FIELDS' | 'UPDATEOWN' | 'UPDATE_SECURE' | 'UPDATE_STATUS' | 'UPDATE_RESULT'